Rename selector. Move Routes to components
