+++ src/App.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react'
 import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'connected-react-router'
-import Routes from './Routes'
+import Routes from './components/Routes'
 // $FlowFixMe
 import 'antd/dist/antd.less'
 import history from './store/history'
